+++ src/Config/Update/Step/AbstractUpdateStep.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Lemberg\Draft\Environment\Config\Install\Step;
+namespace Lemberg\Draft\Environment\Config\Update\Step;
 
 use Composer\Composer;
 use Composer\IO\IOInterface;
@@ -12,9 +12,9 @@ use Lemberg\Draft\Environment\Helper\FileReaderTrait;
 use Lemberg\Draft\Environment\Messanger\MessangerTrait;
 
 /**
- * Default implementation of the installation/uninstall step.
+ * Default implementation of the update step.
  */
-abstract class AbstractInstallStep implements AbstractStepInterface {
+abstract class AbstractUpdateStep implements AbstractStepInterface {
 
   use FileReaderTrait;
   use MessangerTrait;
@@ -32,7 +32,7 @@ abstract class AbstractInstallStep implements AbstractStepInterface {
   /**
    * @var \Lemberg\Draft\Environment\Config\Manager\ManagerInterface
    */
-  protected $configInstallManager;
+  protected $configUpdateManager;
 
   /**
    * {@inheritdoc}
@@ -40,15 +40,8 @@ abstract class AbstractInstallStep implements AbstractStepInterface {
   final public function __construct(Composer $composer, IOInterface $io, ManagerInterface $configManager) {
     $this->composer = $composer;
     $this->io = $io;
-    $this->configInstallManager = $configManager;
+    $this->configUpdateManager = $configManager;
     $this->initFileSystem();
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function getWeight(): int {
-    return 0;
   }
 
 }