+++ src/Messanger/MessangerInterface.php
@@ -1,0 +1,17 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Lemberg\Draft\Environment\Messanger;
+
+/**
+ * Collects and returns messages.
+ */
+interface MessangerInterface {
+
+  /**
+   * @return string
+   */
+  public function getMessages(): string;
+
+}