+++ tests/AppTest.php
@@ -20,7 +20,8 @@ use Composer\Script\Event as ScriptEvent;
 use Composer\Script\ScriptEvents;
 use Lemberg\Draft\Environment\App;
 use Lemberg\Draft\Environment\Config\Config;
-use Lemberg\Draft\Environment\Config\InstallManager;
+use Lemberg\Draft\Environment\Config\Manager\InstallManager;
+use Lemberg\Draft\Environment\Config\Manager\UpdateManager;
 use org\bovigo\vfs\vfsStream;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Filesystem\Filesystem;
@@ -48,9 +49,14 @@ final class AppTest extends TestCase {
   private $app;
 
   /**
-   * @var \Lemberg\Draft\Environment\Config\InstallManager
+   * @var \Lemberg\Draft\Environment\Config\Manager\InstallManager
    */
   private $configInstallManager;
+
+  /**
+   * @var \Lemberg\Draft\Environment\Config\Manager\UpdateManager
+   */
+  private $configUpdateManager;
 
   /**
    *
@@ -97,7 +103,8 @@ final class AppTest extends TestCase {
 
     $config = new Config("$root/source", "$root/target");
     $this->configInstallManager = new InstallManager($this->composer, $this->io, $config);
-    $this->app = new App($this->composer, $this->io, $this->configInstallManager);
+    $this->configUpdateManager = new UpdateManager($this->composer, $this->io, $config);
+    $this->app = new App($this->composer, $this->io, $this->configInstallManager, $this->configUpdateManager);
   }
 
   /**